backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
boto3
langchain
langchain-aws
pydantic